api/mqtt/requirements.txt
paho-mqtt==1.6.1
pytest==7.4.3
pytest-cov==4.1.0 